+++ tests/test_mongo.py
@@ -174,11 +174,12 @@ class TestMongo(MongoTestCase):
                      ('update', update_spec),
                      ('new', True)]))['value']
 
+            def update_worked():
+                replicated = self.mongo_doc.mongo.test.test.find_one({"a": 0})
+                return replicated == updated
 
             # Allow some time for update to propagate
-            time.sleep(2)
-            replicated = self.mongo_doc.mongo.test.test.find_one({"a": 0})
-            self.assertEqual(replicated, updated)
+            assert_soon(update_worked)
 
         # Update by adding a field
         check_update({"$set": {"b": [{"c": 10}, {"d": 11}]}})
@@ -206,8 +207,8 @@ class TestMongo(MongoTestCase):
 
     def test_rollback(self):
         """Tests rollback. We force a rollback by adding a doc, killing the
-            primary, adding another doc, killing the new primary, and then
-            restarting both.
+        primary, adding another doc, killing the new primary, and then
+        restarting both.
         """
         primary_conn = self.repl_set.primary.client()
         self.conn['test']['test'].insert_one({'name': 'paul'})
@@ -228,7 +229,7 @@ class TestMongo(MongoTestCase):
         result_set_1 = list(self._search())
         result_set_2 = self.conn['test']['test'].find_one({'name': 'pauline'})
         self.assertEqual(len(result_set_1), 2)
-        #make sure pauline is there
+        # make sure pauline is there
         for item in result_set_1:
             if item['name'] == 'pauline':
                 self.assertEqual(item['_id'], result_set_2['_id'])